+++ src/dotnet/ToolPackageObtainer/PackageVersion.cs
@@ -3,7 +3,7 @@
 
 using System.IO;
 
-namespace Microsoft.DotNet.ExecutablePackageObtainer
+namespace Microsoft.DotNet.ToolPackageObtainer
 {
     internal class PackageVersion
     {
