+++ src/dotnet/ToolPackageObtainer/ToolConfigurationDeserialization/DotNetCliTool.cs
@@ -1,7 +1,7 @@
 ﻿using System.Diagnostics;
 using System.Xml.Serialization;
 
-namespace Microsoft.DotNet.ExecutablePackageObtainer.ToolConfigurationDeserialization
+namespace Microsoft.DotNet.ToolPackageObtainer.ToolConfigurationDeserialization
 {
     [DebuggerStepThrough]
     [XmlRoot(Namespace = "", IsNullable = false)]
