--- test/Microsoft.DotNet.ExecutablePackageObtainer.Tests/DotnetToolsConfigGolden.xml
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<DotNetCliTool>
-  <Commands>
-    <Command Name="sayhello" EntryPoint="console.dll" Runner="dotnet" />
-  </Commands>
-</DotNetCliTool>
