--- test/Microsoft.DotNet.ExecutablePackageObtainer.Tests/DotnetToolsConfigMissing.xml
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<DotNetCliTool>
-  <Commands>
-    <Command EntryPoint="console.dll" Runner="dotnet" />
-  </Commands>
-</DotNetCliTool>
