--- test/Microsoft.DotNet.ExecutablePackageObtainer.Tests/SampleGlobalTool/DotnetToolsConfig.xml
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<DotNetCliTool>
-  <Commands>
-    <Command Name="demo" EntryPoint="consoledemo.dll" Runner="dotnet" />
-  </Commands>
-</DotNetCliTool>
