--- test/Microsoft.DotNet.ExecutablePackageObtainer.Tests/SampleGlobalTool/includepublish.nuspec
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
-    <metadata>
-        <id>global.tool.console.demo</id>
-        <version>1.0.4</version>
-        <description>test app</description>
-        <authors>testauthor</authors>
-    </metadata>
-    <files>
-        <file src="bin\Release\netcoreapp2.1\publish\*.*" target="tools\netcoreapp2.1\" /> 
-        <file src="DotnetToolsConfig.xml" target="tools\DotnetToolsConfig.xml" /> 
-    </files>
-</package>
