+++ test/Microsoft.DotNet.ToolPackageObtainer.Tests/DotnetToolsConfigMissing.xml
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<DotNetCliTool>
+  <Commands>
+    <Command EntryPoint="console.dll" Runner="dotnet" />
+  </Commands>
+</DotNetCliTool>
