+++ test/Microsoft.DotNet.ToolPackageObtainer.Tests/SampleGlobalTool/DotnetToolsConfig.xml
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<DotNetCliTool>
+  <Commands>
+    <Command Name="demo" EntryPoint="consoledemo.dll" Runner="dotnet" />
+  </Commands>
+</DotNetCliTool>
