+++ test/Microsoft.DotNet.ToolPackageObtainer.Tests/SampleGlobalTool/Program.cs
